changed exports of fn  to match files
