+++ lib/detectImageType.js
@@ -8,7 +8,7 @@ import { specificHandlers } from './types/specificHandlers';
  * @param {function} toAscii - function to transform byte to ascii string
  * @returns {imageType | undefined} - returns image type (as string)
  **/
-const detectType = (view, toAscii) => {
+export default function detectImageType(view, toAscii) {
     const detectionByFirstByte = () => {
         const byte = view.getUint8(0);
         if (byte in firstBytes) {
@@ -37,6 +37,5 @@ const detectType = (view, toAscii) => {
     else {
         return loopThruSpecificHandlers();
     }
-};
-detectType.default = detectType;
-export default detectType;
+}
+;
