+++ lib/index.js
@@ -1,7 +1,7 @@
-import lookup from './lookup';
-import detectType from './detectImageType';
-import viewTiff from './viewTiffImage';
-import { isBigEndian } from './types/tiffHelpers';
+import sizeOf from './sizeOf';
+export { default as detectImageType } from './detectImageType';
+export { default as viewTiffImage } from './viewTiffImage';
+export { default as isTiffBigEndian } from './isTiffBigEndian';
 /**
  * get image size and type from a DataView of buffer
  *
@@ -9,23 +9,15 @@ import { isBigEndian } from './types/tiffHelpers';
  * @param {function} toAscii - function to transform byte to ascii string
  * @returns {
  *  {
- *  width: number,
- *  height: number,
- *  orientation?: number,
- *  type?: string
- * }
+ *    width: number,
+ *    height: number,
+ *    images: []
+ *    type: string
+ *  }
  * } image size
  */
 export const imageSize = (view, toAscii) => {
-    if (typeof toAscii !== 'function') {
-        throw new Error('toAscii is not a callback function');
-    }
-    // detect the file type.. don't rely on the extension
-    const type = detectType(view, toAscii);
-    return lookup(type, view, toAscii);
+    return sizeOf(view, toAscii);
 };
-export const detectImageType = detectType;
-export const isTiffBigEndian = isBigEndian;
-export const viewTiffImage = viewTiff;
-imageSize.default = imageSize;
+imageSize.default = sizeOf;
 export default imageSize;
