+++ lib/types/tiff.js
@@ -1,6 +1,6 @@
-import '../readUInt';
 import toHexadecimal from '../toHexadecimal';
-import { isBigEndian as isTiffBigEndian, getIdfOffset as getTiffIdfOffset, extractIdfEntry as extractTiffIdfEntry, intoResult as intoTiffResult, getNoOfIdfEntries as getTiffNoOfIdfEntries, beginIndexForIdf as beginIndexForTiffIdf, endIndexForIdf as endIndexForTiffIdf, getStrideForIdf as getStrideForTiffIdf, getIdfOffsetLocation as getTiffIdfOffsetLocation, } from './tiffHelpers';
+import isTiffBigEndian from '../isTiffBigEndian';
+import { getIdfOffset as getTiffIdfOffset, extractIdfEntry as extractTiffIdfEntry, intoResult as intoTiffResult, getNoOfIdfEntries as getTiffNoOfIdfEntries, beginIndexForIdf as beginIndexForTiffIdf, endIndexForIdf as endIndexForTiffIdf, getStrideForIdf as getStrideForTiffIdf, getIdfOffsetLocation as getTiffIdfOffsetLocation, } from './tiffHelpers';
 const signatures = [
     // '492049', // currently not supported
     '49492a00',
