+++ lib/types/tiffHelpers.js
@@ -1,17 +1,4 @@
 import { readUInt16, readUInt32 } from '../readUInt';
-// Test if the TIFF is Big Endian or Little Endian
-export function isBigEndian(view, toAscii, offset) {
-    const signature = toAscii(view, offset, 2);
-    if ('II' === signature) {
-        return false;
-    }
-    else if ('MM' === signature) {
-        return true;
-    }
-    else {
-        throw new TypeError(`Tiff endian error - ${signature}`);
-    }
-}
 const TIFF_FILE_DEFAULT = 4;
 export function getIdfOffsetLocation() {
     return TIFF_FILE_DEFAULT;
