+++ package.json
@@ -53,16 +53,13 @@
     "@types/jest": "^26.0.22",
     "@types/node": "^14.14.37",
     "del": "^6.0.0",
-    "esbuild": "^0.9.7",
     "glob": "^7.1.6",
     "jest": "^26.6.3",
     "prettier": "^2.2.1",
     "ts-jest": "^26.5.4",
     "ts-loader": "^8.1.0",
-    "tsconfig-paths-webpack-plugin": "^3.5.1",
     "typescript": "^4.2.4",
     "webpack": "^5.33.2",
-    "webpack-clean": "^1.2.5",
     "webpack-cli": "^4.6.0"
   }
 }
