+++ src/types/tiff.ts
@@ -1,8 +1,7 @@
-import { readUInt16, readUInt32 } from '../readUInt';
 import toHexadecimal from '../toHexadecimal';
 import type { IImage } from './interface';
+import isTiffBigEndian from '../isTiffBigEndian';
 import {
-  isBigEndian as isTiffBigEndian,
   getIdfOffset as getTiffIdfOffset,
   extractIdfEntry as extractTiffIdfEntry,
   intoResult as intoTiffResult,
