+++ types/isTiffBigEndian.d.ts
@@ -1,0 +1,2 @@
+import type { ToAsciiCallback } from './types/interface';
+export default function isTiffBigEndian(view: DataView, toAscii: ToAsciiCallback, offset: number): boolean;
