+++ types/sizeOf.d.ts
@@ -1,8 +1,4 @@
 import type { ISizeCalculationResult, ToAsciiCallback } from './types/interface';
-import sizeOf from './sizeOf';
-export { default as detectImageType } from './detectImageType';
-export { default as viewTiffImage } from './viewTiffImage';
-export { default as isTiffBigEndian } from './isTiffBigEndian';
 /**
  * get image size and type from a DataView of buffer
  *
@@ -17,8 +13,4 @@ export { default as isTiffBigEndian } from './isTiffBigEndian';
  *  }
  * } image size
  */
-export declare const imageSize: {
-    (view: DataView, toAscii: ToAsciiCallback): ISizeCalculationResult;
-    default: typeof sizeOf;
-};
-export default imageSize;
+export default function sizeOf(view: DataView, toAscii: ToAsciiCallback): ISizeCalculationResult;
